Rename to dropdown option intensity.
