+++ src/main/java/swaps/view/DropdownOptionIntensity.java
@@ -1,0 +1,18 @@
+package swaps.view;
+
+public enum DropdownOptionIntensity {
+    HIGH("A lot"),
+    MEDIUM("A moderate amount"),
+    LOW("A little");
+
+    private final String label;
+
+    DropdownOptionIntensity(String label) {
+        this.label = label;
+    }
+
+    @Override
+    public String toString() {
+        return label;
+    }
+}
